Egg hatching and save system updates.
Save system issue when instantiating a prefab at runtime. Was looking for the data persistance interface objects before the tama was in the scene but was fixed when I added the find all to both the load game and autosave functions (to be fixed once I figure out a better solution).